requirements.txt
distlib==0.3.9
filelock==3.17.0
numpy==2.2.3
pillow==11.1.0
platformdirs==4.3.6
setuptools==76.0.0
virtualenv==20.29.2
